Split MockDescriptors inner classes into separate files
