+++ src/native/managed/cdacreader/tests/MockDescriptors/MockDescriptors.Thread.cs
@@ -1,0 +1,155 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+
+namespace Microsoft.Diagnostics.DataContractReader.UnitTests;
+
+internal partial class MockDescriptors
+{
+    public class Thread
+    {
+        const bool UseFunclets = false;
+        private const ulong DefaultAllocationRangeStart = 0x0003_0000;
+        private const ulong DefaultAllocationRangeEnd = 0x0004_0000;
+
+        internal Dictionary<DataType, Target.TypeInfo> Types { get; }
+        internal (string Name, ulong Value, string? Type)[] Globals { get; }
+
+        internal TargetPointer FinalizerThreadAddress { get; }
+        internal TargetPointer GCThreadAddress { get; }
+
+        private readonly MockMemorySpace.Builder _builder;
+        private readonly MockMemorySpace.BumpAllocator _allocator;
+
+        private readonly TargetPointer _threadStoreAddress;
+
+        // Most recently added thread. We update its link to the next thread if another thread is added.
+        private TargetPointer _previousThread = TargetPointer.Null;
+
+        public Thread(MockMemorySpace.Builder builder)
+            : this(builder, (DefaultAllocationRangeStart, DefaultAllocationRangeEnd))
+        { }
+
+        public Thread(MockMemorySpace.Builder builder, (ulong Start, ulong End) allocationRange)
+        {
+            _builder = builder;
+            _allocator = _builder.CreateAllocator(allocationRange.Start, allocationRange.End);
+
+            TargetTestHelpers helpers = builder.TargetTestHelpers;
+
+            Types = GetTypes(helpers);
+
+            // Add thread store and set global to point at it
+            MockMemorySpace.HeapFragment threadStoreGlobal = _allocator.Allocate((ulong)helpers.PointerSize, "[global pointer] ThreadStore");
+            MockMemorySpace.HeapFragment threadStore = _allocator.Allocate(Types[DataType.ThreadStore].Size.Value, "ThreadStore");
+            helpers.WritePointer(threadStoreGlobal.Data, threadStore.Address);
+            _builder.AddHeapFragments([threadStoreGlobal, threadStore]);
+            _threadStoreAddress = threadStore.Address;
+
+            // Add finalizer thread and set global to point at it
+            MockMemorySpace.HeapFragment finalizerThreadGlobal = _allocator.Allocate((ulong)helpers.PointerSize, "[global pointer] Finalizer thread");
+            MockMemorySpace.HeapFragment finalizerThread = _allocator.Allocate(Types[DataType.Thread].Size.Value, "Finalizer thread");
+            helpers.WritePointer(finalizerThreadGlobal.Data, finalizerThread.Address);
+            _builder.AddHeapFragments([finalizerThreadGlobal, finalizerThread]);
+            FinalizerThreadAddress = finalizerThread.Address;
+
+            // Add GC thread and set global to point at it
+            MockMemorySpace.HeapFragment gcThreadGlobal = _allocator.Allocate((ulong)helpers.PointerSize, "[global pointer] GC thread");
+            MockMemorySpace.HeapFragment gcThread = _allocator.Allocate(Types[DataType.Thread].Size.Value, "GC thread");
+            helpers.WritePointer(gcThreadGlobal.Data, gcThread.Address);
+            _builder.AddHeapFragments([gcThreadGlobal, gcThread]);
+            GCThreadAddress = gcThread.Address;
+
+            Globals =
+            [
+                (nameof(Constants.Globals.ThreadStore), threadStoreGlobal.Address, null),
+                (nameof(Constants.Globals.FinalizerThread), finalizerThreadGlobal.Address, null),
+                (nameof(Constants.Globals.GCThread), gcThreadGlobal.Address, null),
+                (nameof(Constants.Globals.FeatureEHFunclets), UseFunclets ? 1 : 0, null),
+            ];
+        }
+
+        private static Dictionary<DataType, Target.TypeInfo> GetTypes(TargetTestHelpers helpers)
+        {
+            TargetTestHelpers.LayoutResult exceptionInfoLayout = helpers.LayoutFields(ExceptionInfoFields);
+            TargetTestHelpers.LayoutResult threadLayout = helpers.LayoutFields(ThreadFields);
+            TargetTestHelpers.LayoutResult threadStoreLayout = helpers.LayoutFields(ThreadStoreFields);
+            return new()
+            {
+                [DataType.ExceptionInfo] = new Target.TypeInfo() { Fields = exceptionInfoLayout.Fields, Size = exceptionInfoLayout.Stride },
+                [DataType.Thread] = new Target.TypeInfo() { Fields = threadLayout.Fields, Size = threadLayout.Stride },
+                [DataType.ThreadStore] = new Target.TypeInfo() { Fields = threadStoreLayout.Fields, Size = threadStoreLayout.Stride },
+            };
+        }
+
+        internal void SetThreadCounts(int threadCount, int unstartedCount, int backgroundCount, int pendingCount, int deadCount)
+        {
+            TargetTestHelpers helpers = _builder.TargetTestHelpers;
+            Target.TypeInfo typeInfo = Types[DataType.ThreadStore];
+            Span<byte> data = _builder.BorrowAddressRange(_threadStoreAddress, (int)typeInfo.Size.Value);
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.ThreadStore.ThreadCount)].Offset),
+                threadCount);
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.ThreadStore.UnstartedCount)].Offset),
+                unstartedCount);
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.ThreadStore.BackgroundCount)].Offset),
+                backgroundCount);
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.ThreadStore.PendingCount)].Offset),
+                pendingCount);
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.ThreadStore.DeadCount)].Offset),
+                deadCount);
+        }
+
+        internal TargetPointer AddThread(uint id, TargetNUInt osId)
+        {
+            TargetTestHelpers helpers = _builder.TargetTestHelpers;
+            Target.TypeInfo typeInfo = Types[DataType.Thread];
+            if (UseFunclets)
+                throw new NotImplementedException("todo for funclets: allocate the ExceptionInfo separately");
+            ulong allocSize = typeInfo.Size.Value + (UseFunclets ? 0 : Types[DataType.ExceptionInfo].Size.Value);
+            MockMemorySpace.HeapFragment thread = _allocator.Allocate(allocSize, UseFunclets ? "Thread" : "Thread and ExceptionInfo");
+            Span<byte> data = thread.Data.AsSpan();
+            helpers.Write(
+                data.Slice(typeInfo.Fields[nameof(Data.Thread.Id)].Offset),
+                id);
+            helpers.WriteNUInt(
+                data.Slice(typeInfo.Fields[nameof(Data.Thread.OSId)].Offset),
+                osId);
+            _builder.AddHeapFragment(thread);
+
+            // Add exception info for the thread
+            // Add exception info for the thread
+            // TODO: [cdac] Handle when UseFunclets is true - see NotImplementedException thrown above
+            TargetPointer exceptionInfoAddress = thread.Address + Types[DataType.ExceptionInfo].Size.Value;
+            helpers.WritePointer(
+                data.Slice(typeInfo.Fields[nameof(Data.Thread.ExceptionTracker)].Offset),
+                exceptionInfoAddress);
+
+            ulong threadLinkOffset = (ulong)typeInfo.Fields[nameof(Data.Thread.LinkNext)].Offset;
+            if (_previousThread != TargetPointer.Null)
+            {
+                // Set the next link for the previously added thread to the newly added one
+                helpers.WritePointer(
+                    _builder.BorrowAddressRange(_previousThread + threadLinkOffset, helpers.PointerSize),
+                    thread.Address + threadLinkOffset);
+            }
+            else
+            {
+                // Set the first thread link in the thread store
+                ulong firstThreadLinkAddr = _threadStoreAddress + (ulong)Types[DataType.ThreadStore].Fields[nameof(Data.ThreadStore.FirstThreadLink)].Offset;
+                helpers.WritePointer(
+                    _builder.BorrowAddressRange(firstThreadLinkAddr, helpers.PointerSize),
+                    thread.Address + threadLinkOffset);
+            }
+
+            _previousThread = thread.Address;
+            return thread.Address;
+        }
+    }
+}
