+++ src/native/managed/cdacreader/tests/MockDescriptors/MockDescriptors.cs
@@ -1,0 +1,126 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.Diagnostics.DataContractReader.UnitTests;
+
+internal partial class MockDescriptors
+{
+    private static readonly (string Name, DataType Type)[] MethodTableFields = new[]
+    {
+        (nameof(Data.MethodTable.MTFlags), DataType.uint32),
+        (nameof(Data.MethodTable.BaseSize), DataType.uint32),
+        (nameof(Data.MethodTable.MTFlags2), DataType.uint32),
+        (nameof(Data.MethodTable.EEClassOrCanonMT), DataType.nuint),
+        (nameof(Data.MethodTable.Module), DataType.pointer),
+        (nameof(Data.MethodTable.ParentMethodTable), DataType.pointer),
+        (nameof(Data.MethodTable.NumInterfaces), DataType.uint16),
+        (nameof(Data.MethodTable.NumVirtuals), DataType.uint16),
+        (nameof(Data.MethodTable.PerInstInfo), DataType.pointer),
+        (nameof(Data.MethodTable.AuxiliaryData), DataType.pointer),
+    };
+
+    private static readonly (string Name, DataType Type)[] EEClassFields = new[]
+    {
+        (nameof(Data.EEClass.MethodTable), DataType.pointer),
+        (nameof(Data.EEClass.CorTypeAttr), DataType.uint32),
+        (nameof(Data.EEClass.NumMethods), DataType.uint16),
+        (nameof(Data.EEClass.InternalCorElementType), DataType.uint8),
+        (nameof(Data.EEClass.NumNonVirtualSlots), DataType.uint16),
+    };
+
+    private static readonly (string Name, DataType Type)[] MethodTableAuxiliaryDataFields = new[]
+    {
+        (nameof(Data.MethodTableAuxiliaryData.LoaderModule), DataType.pointer),
+        (nameof(Data.MethodTableAuxiliaryData.OffsetToNonVirtualSlots), DataType.int16),
+    };
+
+    private static readonly (string Name, DataType Type)[] ArrayClassFields = new[]
+    {
+        (nameof(Data.ArrayClass.Rank), DataType.uint8),
+    };
+
+    private static readonly (string Name, DataType Type)[] ObjectFields = new[]
+    {
+        ("m_pMethTab", DataType.pointer),
+    };
+
+    private static readonly (string Name, DataType Type)[] StringFields = new[]
+    {
+        ("m_StringLength", DataType.uint32),
+        ("m_FirstChar", DataType.uint16),
+    };
+
+    private static readonly (string Name, DataType Type)[] ArrayFields = new[]
+    {
+        ("m_NumComponents", DataType.uint32),
+    };
+
+    private static readonly (string Name, DataType Type)[] SyncTableEntryFields = new[]
+    {
+        (nameof(Data.SyncTableEntry.SyncBlock), DataType.pointer),
+    };
+
+    private static readonly (string Name, DataType Type)[] SyncBlockFields = new[]
+    {
+        (nameof(Data.SyncBlock.InteropInfo), DataType.pointer),
+    };
+
+    private static readonly (string Name, DataType Type)[] InteropSyncBlockFields = new[]
+    {
+        (nameof(Data.InteropSyncBlockInfo.RCW), DataType.pointer),
+        (nameof(Data.InteropSyncBlockInfo.CCW), DataType.pointer),
+    };
+
+    private static readonly (string, DataType)[] ModuleFields =
+    [
+        (nameof(Data.Module.Assembly), DataType.pointer),
+        (nameof(Data.Module.Flags), DataType.uint32),
+        (nameof(Data.Module.Base), DataType.pointer),
+        (nameof(Data.Module.LoaderAllocator), DataType.pointer),
+        (nameof(Data.Module.ThunkHeap), DataType.pointer),
+        (nameof(Data.Module.DynamicMetadata), DataType.pointer),
+        (nameof(Data.Module.Path), DataType.pointer),
+        (nameof(Data.Module.FileName), DataType.pointer),
+        (nameof(Data.Module.FieldDefToDescMap), DataType.pointer),
+        (nameof(Data.Module.ManifestModuleReferencesMap), DataType.pointer),
+        (nameof(Data.Module.MemberRefToDescMap), DataType.pointer),
+        (nameof(Data.Module.MethodDefToDescMap), DataType.pointer),
+        (nameof(Data.Module.TypeDefToMethodTableMap), DataType.pointer),
+        (nameof(Data.Module.TypeRefToMethodTableMap), DataType.pointer),
+        (nameof(Data.Module.MethodDefToILCodeVersioningStateMap), DataType.pointer),
+    ];
+
+    private static readonly (string, DataType)[] AssemblyFields =
+    [
+        (nameof(Data.Assembly.IsCollectible), DataType.uint8),
+    ];
+    private static readonly (string, DataType)[] ExceptionInfoFields =
+    [
+        (nameof(Data.ExceptionInfo.PreviousNestedInfo), DataType.pointer),
+        (nameof(Data.ExceptionInfo.ThrownObject), DataType.pointer),
+    ];
+
+    private static readonly (string, DataType)[] ThreadFields =
+    [
+        (nameof(Data.Thread.Id), DataType.uint32),
+        (nameof(Data.Thread.OSId), DataType.nuint),
+        (nameof(Data.Thread.State), DataType.uint32),
+        (nameof(Data.Thread.PreemptiveGCDisabled), DataType.uint32),
+        (nameof(Data.Thread.RuntimeThreadLocals), DataType.pointer),
+        (nameof(Data.Thread.Frame), DataType.pointer),
+        (nameof(Data.Thread.TEB), DataType.pointer),
+        (nameof(Data.Thread.LastThrownObject), DataType.pointer),
+        (nameof(Data.Thread.LinkNext), DataType.pointer),
+        (nameof(Data.Thread.ExceptionTracker), DataType.pointer),
+    ];
+
+    private static readonly (string, DataType)[] ThreadStoreFields =
+    [
+        (nameof(Data.ThreadStore.ThreadCount), DataType.uint32),
+        (nameof(Data.ThreadStore.FirstThreadLink), DataType.pointer),
+        (nameof(Data.ThreadStore.UnstartedCount), DataType.uint32),
+        (nameof(Data.ThreadStore.BackgroundCount), DataType.uint32),
+        (nameof(Data.ThreadStore.PendingCount), DataType.uint32),
+        (nameof(Data.ThreadStore.DeadCount), DataType.uint32),
+    ];
+}
